Integrate Viewport handling in Engine and SkiaRenderer; update rendering logic to utilize scroll functionality and adjust layout calculations based on Viewport dimensions. Refactor MainWindow to correctly initialize siteContentWidget and manage screen dimensions.

+++ include/System/Screen.hpp
@@ -2,6 +2,6 @@
 
 namespace Screen
 {
-    inline int width = 0;
-    inline int height = 0;
+    inline float width = 0;
+    inline float height = 0;
 }
